fix(timing): Promises that finish late are not supported correctly
For example this is a very common one (happens in the aurelia-validation examples, currently):
User types "A", AJAX call to validate "A" is triggered.
User types backslash, '' is validated as valid because this is not a required field
AJAX call to validate "A" completes, field is marked as invalid.
=> result: '' is considered invalid, even though it was not a required field

Fixes #27