feat(onValidate): callback for async validation on the entire subject
Using the new .onValidate(), you can register a callback that is called when the entire subject is validated.  Your function should return an object (or a promise that resolves to an object) that has properties matching each validation property you want to modify, for example:
{
  firstName : true, //is valid
  lastName : null, //is valid
  middleName : '', //is valid
  addressLine1 : false, //is not valid
  addressLine2 : 'should be in Antarctica' //is not valid with custom message
}
You're not required to return each property in the validation group, and you're not required to return the same properties every time.

Fixes #25