doc(decorators): introduced the @ensure decorator
Having a fluent API allows you to keep your code that sets up the validation logic tightly together. However, some people prefer to keep each aspect of their validation close to the property on which it reflects.  You can do this by adding a decorator called @ensure on each property that requires validation.  
The @ensure decorator takes a callback function that will be called when validation is actually set up.