Fix computedFrom / dependencies not working
Fix for #73

Dependencies do not trigger a validation if the target is already marked dirty (basically anytime after first dependency change or if the property was already modified) and if the property's own value does not change. 

Of course, a dependency change should always trigger validation (hence `forceExecution: true`).

I always changed back `forceDirty` to `false` because a dependency change should not mark a property dirty. Imaging you have `fromDate` and `toDate` with a dependency `.ensure('toDate', config => config.computedFrom(['fromDate']))` and rules such as if `fromDate` is not empty `toDate` must not be empty and must be greater. It is not a good UX that as soon as the user types in `fromDate` the validation on `toDate` appears in UI. It should only appear after the user attempts to change `toDate` or tries to submit the form.