<feat>(ValidationGroup) add ValidationGroup.onResultPropertyChanged
The existing ValidationGroup.onValidate callbacks are only triggered on a
manual ValidationGroup.validate() call, but the setting and unsetting of
validity on any added rules is still being performed - based on the
binding model - in the background.
This pull request suggests adding an extra
ValidationGroup.onResultPropertyChanged(callback) function, which will
trigger the callback on each occasion that a ValidationResultProperty's
onValidate callbacks are triggered.

Closes #214