feat(validation-dependency): add support for validation binding dependency
This enables the `ValidationController` to keep track of its bindings by
property name, then validate bindings inspecting their rules for
dependencies and if found validate those bindings too.

This also introduces `ValidationRule` and `RulesContainer` as interfaces to
define abstract validation rules and an object to contain them; plus,
a metadata-key to set a `RulesContainer` on a type metadata.