Internet Explorer does not support closest
I did not test the code before pushing the button. I made some changes that are not in my code. I also added the extra check to see if we traversed back to BODY.

I don't see it in this implementation but I also had to change `message.remove()` to `message.parentNode.removeChild(message);`