IE does not support closest
I still argue for the sake of simplicity not to have the overhead and burden of using a polyfill. I also had to make change to the code to use `element` instead of `target`. I tested this change as well. The thinking is that I can either copy and paste what is given or I have to install and configure my app to use a polyfill for a method that is only used in two places. A 5 second fix or 5 minute fix. This change also follows the belief of not relying on other peoples code.