Clean up Custom Rules
Added value.trim() === '' test.
Corrected \${$config.max}
There was a bit more involved because the value needs to be converted to the respected type then a test needs to be done on the converted type and not the string value.
I tried the value === '' but then I started to think if the user enters a space.
I still think beyond this clean up there need need more to explain the `null` and `undefined` and `value.trim() === ''` stuff. The parallel testing without bail.