Add Validator parameter to ValidationController.createForCurrentScope
It is not currently possible to provide a Validater to a specific
instance of ValidationController, at least, not without using injection.

ValidationController.createForCurrentScope has been provided to enable
us to instantiate ValidationController without the use of injection, but
provides no means of supplying a Validator for the controller instance.

I have a use case where the ValidationController needs to be
instantiated without the use of injection because the code is in a base
class where, due to javascript/typescript limitations, injection cannot
be used.

Note: I'm not altering the documentation because it ties into:

1) documentation about configuring the plugin, which I believe is
missing altogether
2) other means of assigning a Validator to a controller (through
injection) that I understand are being added but may yet have been
documented.

Thus I think someone with more broad knowledge of all of this should
undertake the task of documenting the whole thing together.