feat(PropertyAccessorParser): Remove PropertyAccessorParser in favor of the keyof operator.
Calls to .ensure() allowed both a string and a function as argument. The
function was never actually executed, but parsed to get the string
property name. The same effect can be achieved by using the keyof
operator. In this case, only strings can be provided that are a property
of the given class.

Initial version, requiring the programmer to explicitly state the object
class each time .ensure() is executed. This is because .ensure() is a
static function, so the type cannot be determined yet.

Fixes #485.